// frontend/src/api/generated/models/CreateProject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inundated
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateProject
 */
export interface CreateProject {
    /**
     * 
     * @type {string}
     * @memberof CreateProject
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProject
     */
    color: string;
    /**
     * 
     * @type {number}
     * @memberof CreateProject
     */
    timeBudgetHours?: number;
    /**
     * 
     * @type {Set<string>}
     * @memberof CreateProject
     */
    tagIds?: Set<string>;
}

/**
 * Check if a given object implements the CreateProject interface.
 */
export function instanceOfCreateProject(value: object): value is CreateProject {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('color' in value) || value['color'] === undefined) return false;
    return true;
}

export function CreateProjectFromJSON(json: any): CreateProject {
    return CreateProjectFromJSONTyped(json, false);
}

export function CreateProjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateProject {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'color': json['color'],
        'timeBudgetHours': json['timeBudgetHours'] == null ? undefined : json['timeBudgetHours'],
        'tagIds': json['tagIds'] == null ? undefined : new Set(json['tagIds']),
    };
}

export function CreateProjectToJSON(json: any): CreateProject {
    return CreateProjectToJSONTyped(json, false);
}

export function CreateProjectToJSONTyped(value?: CreateProject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'color': value['color'],
        'timeBudgetHours': value['timeBudgetHours'],
        'tagIds': value['tagIds'] == null ? undefined : Array.from(value['tagIds'] as Set<any>),
    };
}